Selection Sort

+++ 2-selection_sort.c
@@ -8,7 +8,7 @@
  */
 void selection_sort(int *array, size_t size)
 {
-	int i, j, min_idx, tmp;
+	size_t i, j, min_idx, tmp;
 
 	for (i = 0; i < size; i++)
 	{
